Moved helper functions to helpers.py
