+++ homeassistant/components/lcn/switch.py
@@ -5,8 +5,8 @@ from homeassistant.components.switch import SwitchDevice
 from homeassistant.const import CONF_ADDRESS
 
 from . import LcnDevice
-from .const import (
-    CONF_CONNECTIONS, CONF_OUTPUT, DATA_LCN, OUTPUT_PORTS, get_connection)
+from .const import CONF_CONNECTIONS, CONF_OUTPUT, DATA_LCN, OUTPUT_PORTS
+from .helpers import get_connection
 
 
 async def async_setup_platform(hass, hass_config, async_add_entities,
